requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
ipykernel
pandas
numpy
matplotlib
scikit-learn
requests
tqdm
